refactor(blueprint-bin): split main.rs god object into peripheral modules
main.rs (1111 LOC) -> keeps the header, the tokio main() entrypoint, and
setup_log; extracts the peripheral machinery: workflow_status (status/list/
detail HTTP handlers + router), consumer (LoggingHeartbeatConsumer + the
reconciling Tangle job-result consumer), bootstrap (env/heartbeat config +
chain-vs-host capacity check), and the unit tests. Pure code motion, no
behavior change; main.rs now 490 LOC. main() itself stays a single large
entrypoint fn — decomposing it is a separate behavior-risk change.
