+++ ai-agent-sandbox-blueprint-bin/src/main.rs
@@ -34,132 +34,13 @@ use blueprint_qos::heartbeat::{HeartbeatConfig, HeartbeatConsumer};
 #[cfg(feature = "qos")]
 use blueprint_qos::metrics::MetricsConfig;
 
-fn workflow_status_error(error: WorkflowStatusError) -> (StatusCode, Json<serde_json::Value>) {
-    let status = match &error {
-        WorkflowStatusError::NotFound(_) => StatusCode::NOT_FOUND,
-        WorkflowStatusError::Forbidden(_) => StatusCode::FORBIDDEN,
-        WorkflowStatusError::Internal(_) => StatusCode::INTERNAL_SERVER_ERROR,
-    };
+mod bootstrap;
+mod consumer;
+mod workflow_status;
 
-    (
-        status,
-        Json(serde_json::json!({
-            "error": error.message(),
-        })),
-    )
-}
-
-async fn workflow_status_handler(
-    sandbox_runtime::session_auth::SessionAuth(caller): sandbox_runtime::session_auth::SessionAuth,
-    Path(workflow_id): Path<u64>,
-) -> Result<
-    Json<ai_agent_sandbox_blueprint_lib::workflows::WorkflowRuntimeStatus>,
-    (StatusCode, Json<serde_json::Value>),
-> {
-    workflow_runtime_status_for_owner(workflow_id, caller.as_str())
-        .map(Json)
-        .map_err(workflow_status_error)
-}
-
-async fn workflow_list_handler(
-    sandbox_runtime::session_auth::SessionAuth(caller): sandbox_runtime::session_auth::SessionAuth,
-) -> Result<Json<serde_json::Value>, (StatusCode, Json<serde_json::Value>)> {
-    ai_agent_sandbox_blueprint_lib::workflows::list_workflows_for_owner(caller.as_str())
-        .map(|workflows| {
-            Json(serde_json::json!({
-                "workflows": workflows
-                    .into_iter()
-                    .map(|workflow| serde_json::json!({
-                        "scope": "sandbox",
-                        "workflowId": workflow.workflow_id,
-                        "name": workflow.name,
-                        "triggerType": workflow.trigger_type,
-                        "triggerConfig": workflow.trigger_config,
-                        "targetKind": workflow.target_kind,
-                        "targetSandboxId": workflow.target_sandbox_id,
-                        "targetServiceId": workflow.target_service_id,
-                        "active": workflow.active,
-                        "targetStatus": workflow.target_status,
-                        "runnable": workflow.runnable,
-                        "running": workflow.running,
-                        "lastRunAt": workflow.last_run_at,
-                        "nextRunAt": workflow.next_run_at,
-                        "latestExecution": workflow.latest_execution,
-                    }))
-                    .collect::<Vec<_>>(),
-            }))
-        })
-        .map_err(workflow_status_error)
-}
-
-async fn workflow_detail_handler(
-    sandbox_runtime::session_auth::SessionAuth(caller): sandbox_runtime::session_auth::SessionAuth,
-    Path(workflow_id): Path<u64>,
-) -> Result<Json<serde_json::Value>, (StatusCode, Json<serde_json::Value>)> {
-    ai_agent_sandbox_blueprint_lib::workflows::workflow_detail_for_owner(
-        workflow_id,
-        caller.as_str(),
-    )
-    .map(|workflow| {
-        Json(serde_json::json!({
-            "scope": "sandbox",
-            "workflowId": workflow.workflow_id,
-            "name": workflow.name,
-            "workflowJson": workflow.workflow_json,
-            "triggerType": workflow.trigger_type,
-            "triggerConfig": workflow.trigger_config,
-            "sandboxConfigJson": workflow.sandbox_config_json,
-            "targetKind": workflow.target_kind,
-            "targetSandboxId": workflow.target_sandbox_id,
-            "targetServiceId": workflow.target_service_id,
-            "active": workflow.active,
-            "targetStatus": workflow.target_status,
-            "runnable": workflow.runnable,
-            "running": workflow.running,
-            "lastRunAt": workflow.last_run_at,
-            "nextRunAt": workflow.next_run_at,
-            "latestExecution": workflow.latest_execution,
-        }))
-    })
-    .map_err(workflow_status_error)
-}
-
-fn workflow_status_router() -> HttpRouter {
-    HttpRouter::new()
-        .route("/api/workflows", get(workflow_list_handler))
-        .route("/api/workflows/{workflow_id}", get(workflow_status_handler))
-        .route(
-            "/api/workflows/{workflow_id}/detail",
-            get(workflow_detail_handler),
-        )
-}
-
-/// Logging heartbeat consumer that records heartbeat submissions.
-///
-/// The actual on-chain submission is handled internally by `HeartbeatService`
-/// via ECDSA signing + `submitHeartbeat` contract call. This consumer provides
-/// a hook for blueprint-level logging/monitoring of heartbeat events.
-#[cfg(feature = "qos")]
-#[derive(Clone)]
-struct LoggingHeartbeatConsumer;
-
-#[cfg(feature = "qos")]
-impl HeartbeatConsumer for LoggingHeartbeatConsumer {
-    fn send_heartbeat(
-        &self,
-        status: &blueprint_qos::heartbeat::HeartbeatStatus,
-    ) -> std::pin::Pin<
-        Box<dyn std::future::Future<Output = blueprint_qos::error::Result<()>> + Send + 'static>,
-    > {
-        let service_id = status.service_id;
-        let status_code = status.status_code;
-        let ts = status.timestamp;
-        Box::pin(async move {
-            info!("Heartbeat sent: service={service_id} status={status_code} ts={ts}");
-            Ok(())
-        })
-    }
-}
+use bootstrap::*;
+use consumer::*;
+use workflow_status::*;
 
 #[tokio::main]
 #[allow(clippy::result_large_err)]
@@ -594,82 +475,6 @@ async fn main() -> Result<(), blueprint_sdk::Error> {
     Ok(())
 }
 
-/// Parse a u64 from the first env var that's set in `keys`. Logs a warning
-/// and returns `None` if a value is set but doesn't parse — so operators
-/// see misconfiguration in observability instead of features silently
-/// disabling.
-#[cfg(feature = "qos")]
-fn parse_required_u64_env(keys: &[&str]) -> Option<u64> {
-    for key in keys {
-        match std::env::var(key) {
-            Ok(raw) => match raw.parse::<u64>() {
-                Ok(v) => return Some(v),
-                Err(e) => {
-                    warn!(
-                        env = key,
-                        value = %raw,
-                        err = %e,
-                        "env var is set but not a valid u64; falling back to next key"
-                    );
-                }
-            },
-            Err(_) => continue,
-        }
-    }
-    None
-}
-
-/// Build heartbeat config from environment variables.
-///
-/// Required env vars:
-///   - `SERVICE_ID` or `TANGLE_SERVICE_ID` — the service instance ID
-///   - `BLUEPRINT_ID` or `TANGLE_BLUEPRINT_ID` — the blueprint ID
-///   - `STATUS_REGISTRY_ADDRESS` — the OperatorStatusRegistry contract address
-///
-/// Optional:
-///   - `HEARTBEAT_INTERVAL_SECS` — heartbeat interval (default: 120)
-///   - `HEARTBEAT_MAX_MISSED` — max missed beats before slashing (default: 3)
-#[cfg(feature = "qos")]
-fn build_heartbeat_config() -> Option<HeartbeatConfig> {
-    use std::str::FromStr;
-
-    let service_id: u64 = parse_required_u64_env(&["SERVICE_ID", "TANGLE_SERVICE_ID"])?;
-    let blueprint_id: u64 = parse_required_u64_env(&["BLUEPRINT_ID", "TANGLE_BLUEPRINT_ID"])?;
-
-    let registry_addr_str = std::env::var("STATUS_REGISTRY_ADDRESS").ok()?;
-    let status_registry_address =
-        match blueprint_sdk::alloy::primitives::Address::from_str(&registry_addr_str) {
-            Ok(addr) => addr,
-            Err(e) => {
-                warn!(
-                    value = %registry_addr_str,
-                    err = %e,
-                    "STATUS_REGISTRY_ADDRESS is set but not a valid EVM address; heartbeat disabled"
-                );
-                return None;
-            }
-        };
-
-    let interval_secs: u64 = std::env::var("HEARTBEAT_INTERVAL_SECS")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(120);
-
-    let max_missed: u32 = std::env::var("HEARTBEAT_MAX_MISSED")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(3);
-
-    Some(HeartbeatConfig {
-        interval_secs,
-        jitter_percent: 10,
-        service_id,
-        blueprint_id,
-        max_missed_heartbeats: max_missed,
-        status_registry_address,
-    })
-}
-
 fn setup_log() {
     use tracing_subscriber::prelude::*;
     use tracing_subscriber::{EnvFilter, fmt};
@@ -681,431 +486,5 @@ fn setup_log() {
     {}
 }
 
-struct DerivedJobResult {
-    service_id: u64,
-    call_id: u64,
-    output: blueprint_sdk::alloy::primitives::Bytes,
-}
-
-enum ConsumerState {
-    WaitingForResult,
-    ProcessingSubmission(
-        Pin<Box<dyn std::future::Future<Output = Result<(), ReconciledConsumerError>> + Send>>,
-    ),
-}
-
-impl ConsumerState {
-    fn is_waiting(&self) -> bool {
-        matches!(self, Self::WaitingForResult)
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-enum ReconciledConsumerError {
-    #[error("Invalid metadata: {0}")]
-    InvalidMetadata(&'static str),
-    #[error("Transaction error: {0}")]
-    Transaction(String),
-}
-
-struct ReconciledTangleConsumer {
-    client: Arc<TangleClient>,
-    buffer: Mutex<VecDeque<DerivedJobResult>>,
-    state: Mutex<ConsumerState>,
-}
-
-impl ReconciledTangleConsumer {
-    fn new(client: TangleClient) -> Self {
-        Self {
-            client: Arc::new(client),
-            buffer: Mutex::new(VecDeque::new()),
-            state: Mutex::new(ConsumerState::WaitingForResult),
-        }
-    }
-}
-
-impl Sink<blueprint_sdk::JobResult> for ReconciledTangleConsumer {
-    type Error = BoxError;
-
-    fn poll_ready(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        Poll::Ready(Ok(()))
-    }
-
-    fn start_send(self: Pin<&mut Self>, item: blueprint_sdk::JobResult) -> Result<(), Self::Error> {
-        let blueprint_sdk::JobResult::Ok { head, body } = &item else {
-            blueprint_sdk::trace!(target: "tangle-consumer", "Discarding job result with error");
-            return Ok(());
-        };
-
-        let (Some(call_id_raw), Some(service_id_raw)) = (
-            head.metadata.get(CallId::METADATA_KEY),
-            head.metadata.get(ServiceId::METADATA_KEY),
-        ) else {
-            blueprint_sdk::trace!(
-                target: "tangle-consumer",
-                "Discarding job result with missing metadata"
-            );
-            return Ok(());
-        };
-
-        let call_id: u64 = call_id_raw
-            .try_into()
-            .map_err(|_| ReconciledConsumerError::InvalidMetadata("call_id"))?;
-        let service_id: u64 = service_id_raw
-            .try_into()
-            .map_err(|_| ReconciledConsumerError::InvalidMetadata("service_id"))?;
-
-        self.get_mut()
-            .buffer
-            .lock()
-            .unwrap()
-            .push_back(DerivedJobResult {
-                service_id,
-                call_id,
-                output: blueprint_sdk::alloy::primitives::Bytes::copy_from_slice(body),
-            });
-        Ok(())
-    }
-
-    fn poll_flush(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        let consumer = self.get_mut();
-        let mut state = consumer.state.lock().unwrap();
-
-        {
-            let buffer = consumer.buffer.lock().unwrap();
-            if buffer.is_empty() && state.is_waiting() {
-                return Poll::Ready(Ok(()));
-            }
-        }
-
-        loop {
-            match &mut *state {
-                ConsumerState::WaitingForResult => {
-                    let next = {
-                        let mut buffer = consumer.buffer.lock().unwrap();
-                        buffer.pop_front()
-                    };
-
-                    let Some(DerivedJobResult {
-                        service_id,
-                        call_id,
-                        output,
-                    }) = next
-                    else {
-                        return Poll::Ready(Ok(()));
-                    };
-
-                    let client = Arc::clone(&consumer.client);
-                    let fut = Box::pin(async move {
-                        submit_result_and_reconcile(client, service_id, call_id, output).await
-                    });
-                    *state = ConsumerState::ProcessingSubmission(fut);
-                }
-                ConsumerState::ProcessingSubmission(future) => match future.as_mut().poll(cx) {
-                    Poll::Ready(Ok(())) => {
-                        *state = ConsumerState::WaitingForResult;
-                    }
-                    Poll::Ready(Err(err)) => {
-                        *state = ConsumerState::WaitingForResult;
-                        return Poll::Ready(Err(err.into()));
-                    }
-                    Poll::Pending => return Poll::Pending,
-                },
-            }
-        }
-    }
-
-    fn poll_close(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        let buffer = self.buffer.lock().unwrap();
-        if buffer.is_empty() {
-            Poll::Ready(Ok(()))
-        } else {
-            Poll::Pending
-        }
-    }
-}
-
-async fn submit_result_and_reconcile(
-    client: Arc<TangleClient>,
-    service_id: u64,
-    call_id: u64,
-    output: blueprint_sdk::alloy::primitives::Bytes,
-) -> Result<(), ReconciledConsumerError> {
-    if client.config.dry_run {
-        info!(
-            "Dry run enabled; skipping on-chain result submission for service {service_id} call {call_id}"
-        );
-        return Ok(());
-    }
-
-    match client
-        .submit_result(service_id, call_id, output.clone())
-        .await
-    {
-        Ok(result) if result.success => {
-            reconcile_workflows(&client, service_id).await;
-            Ok(())
-        }
-        Ok(result) => Err(ReconciledConsumerError::Transaction(format!(
-            "Transaction reverted for service {service_id} call {call_id}: tx_hash={:?}",
-            result.tx_hash
-        ))),
-        Err(err) if is_job_already_completed(&err.to_string()) => {
-            warn!(
-                "Result for service {service_id} call {call_id} was already completed; treating replay as idempotent"
-            );
-            reconcile_workflows(&client, service_id).await;
-            Ok(())
-        }
-        Err(err)
-            if replay_error_is_already_materialized(
-                &client,
-                service_id,
-                call_id,
-                &output,
-                &err.to_string(),
-            )
-            .await =>
-        {
-            warn!(
-                "Result for service {service_id} call {call_id} is already reflected on-chain; treating replay as idempotent"
-            );
-            reconcile_workflows(&client, service_id).await;
-            Ok(())
-        }
-        Err(err) => Err(ReconciledConsumerError::Transaction(format!(
-            "Failed to submit result for service {service_id} call {call_id}: {err}"
-        ))),
-    }
-}
-
-async fn reconcile_workflows(client: &TangleClient, service_id: u64) {
-    if let Err(err) = bootstrap_workflows_from_chain(client, service_id).await {
-        warn!("Failed to reconcile workflows from chain for service {service_id}: {err}");
-    }
-}
-
-fn is_job_already_completed(error: &str) -> bool {
-    error.contains("JobAlreadyCompleted") || error.contains("already completed")
-}
-
-async fn replay_error_is_already_materialized(
-    client: &TangleClient,
-    service_id: u64,
-    call_id: u64,
-    output: &blueprint_sdk::alloy::primitives::Bytes,
-    error: &str,
-) -> bool {
-    if !error.contains("execution reverted") {
-        return false;
-    }
-
-    if bootstrap_workflows_from_chain(client, service_id)
-        .await
-        .is_err()
-    {
-        return false;
-    }
-
-    let workflow_for_call_id = workflows()
-        .ok()
-        .and_then(|store| store.get(&workflow_key(call_id)).ok())
-        .flatten();
-
-    let payload = JsonResponse::abi_decode(output.as_ref())
-        .ok()
-        .and_then(|response| serde_json::from_str::<Value>(&response.json).ok());
-
-    if let Some(payload) = payload.as_ref() {
-        let Some(workflow_id) = payload.get("workflowId").and_then(Value::as_u64) else {
-            return false;
-        };
-
-        let workflow = workflows()
-            .ok()
-            .and_then(|store| store.get(&workflow_key(workflow_id)).ok())
-            .flatten();
-
-        if workflow_replay_matches_store(call_id, payload, workflow.as_ref()) {
-            return true;
-        }
-    }
-
-    if let Ok(create_output) = SandboxCreateOutput::abi_decode(output.as_ref())
-        && ai_agent_sandbox_blueprint_lib::runtime::get_sandbox_by_id(&create_output.sandboxId)
-            .is_ok()
-    {
-        return true;
-    }
-
-    // Workflow IDs are derived from the create call ID. If a replayed
-    // `workflow_create` result arrives before we can decode its body cleanly,
-    // an active workflow keyed by the same call ID is still enough evidence
-    // that the original result has already been materialized on-chain.
-    workflow_for_call_id
-        .as_ref()
-        .is_some_and(|entry| entry.active)
-}
-
-fn workflow_replay_matches_store(
-    call_id: u64,
-    payload: &Value,
-    workflow: Option<&WorkflowEntry>,
-) -> bool {
-    let Some(workflow_id) = payload.get("workflowId").and_then(Value::as_u64) else {
-        return false;
-    };
-
-    match payload.get("status").and_then(Value::as_str) {
-        Some("canceled") => workflow.is_none(),
-        Some("active") => workflow.as_ref().is_some(),
-        _ if payload.get("task").is_some() => workflow.as_ref().is_some(),
-        _ => workflow_id == call_id && workflow.as_ref().is_some(),
-    }
-}
-
-/// Cross-check on-chain capacity vs the host admission cap.
-///
-/// `OPERATOR_MAX_CAPACITY` is what this operator registers on-chain (the
-/// chain may assign that many sandboxes); `SANDBOX_MAX_COUNT` is what the
-/// host runtime will actually admit. Registering more than the host admits
-/// guarantees rejected work, so startup fails when both are set and
-/// chain > host. `SANDBOX_MAX_COUNT=0` (uncapped host) always passes.
-/// Unparseable values are ignored here for parity with their consumers:
-/// registration skips an unparseable `OPERATOR_MAX_CAPACITY` and the runtime
-/// substitutes its default for an unparseable `SANDBOX_MAX_COUNT`.
-fn validate_chain_vs_host_capacity(
-    operator_max_capacity: Option<&str>,
-    sandbox_max_count: Option<&str>,
-) -> Result<(), String> {
-    let (Some(capacity_raw), Some(max_count_raw)) = (operator_max_capacity, sandbox_max_count)
-    else {
-        return Ok(());
-    };
-    let (Ok(capacity), Ok(max_count)) = (
-        capacity_raw.trim().parse::<u32>(),
-        max_count_raw.trim().parse::<usize>(),
-    ) else {
-        return Ok(());
-    };
-    if max_count == 0 || capacity as usize <= max_count {
-        return Ok(());
-    }
-    Err(format!(
-        "OPERATOR_MAX_CAPACITY={capacity} exceeds SANDBOX_MAX_COUNT={max_count}: the chain \
-         would assign up to {capacity} sandboxes but this host rejects creations beyond \
-         {max_count}. Lower OPERATOR_MAX_CAPACITY or raise SANDBOX_MAX_COUNT."
-    ))
-}
-
 #[cfg(test)]
-mod tests {
-    use super::{WorkflowEntry, validate_chain_vs_host_capacity, workflow_replay_matches_store};
-    use serde_json::json;
-
-    fn active_workflow(id: u64) -> WorkflowEntry {
-        WorkflowEntry {
-            id,
-            name: "workflow-qa".into(),
-            workflow_json: "{}".into(),
-            trigger_type: "cron".into(),
-            trigger_config: "0 * * * * *".into(),
-            sandbox_config_json: "{}".into(),
-            target_kind: 0,
-            target_sandbox_id: String::new(),
-            target_service_id: 0,
-            active: true,
-            next_run_at: None,
-            last_run_at: None,
-            owner: String::new(),
-        }
-    }
-
-    #[test]
-    fn create_replay_matches_existing_active_workflow() {
-        let payload = json!({
-            "status": "active",
-            "workflowId": 7
-        });
-
-        assert!(workflow_replay_matches_store(
-            7,
-            &payload,
-            Some(&active_workflow(7))
-        ));
-    }
-
-    #[test]
-    fn trigger_replay_matches_existing_workflow_even_if_inactive_bit_isnt_rechecked() {
-        let payload = json!({
-            "status": "active",
-            "workflowId": 9,
-            "task": {
-                "success": true
-            }
-        });
-
-        assert!(workflow_replay_matches_store(
-            12,
-            &payload,
-            Some(&active_workflow(9))
-        ));
-    }
-
-    #[test]
-    fn canceled_replay_only_matches_when_active_store_entry_is_absent() {
-        let payload = json!({
-            "status": "canceled",
-            "workflowId": 11
-        });
-
-        assert!(workflow_replay_matches_store(15, &payload, None));
-        assert!(!workflow_replay_matches_store(
-            15,
-            &payload,
-            Some(&active_workflow(11))
-        ));
-    }
-
-    // ── chain-vs-host capacity cross-check ─────────────────────────────
-
-    #[test]
-    fn capacity_cross_check_fails_when_chain_exceeds_host() {
-        let err = validate_chain_vs_host_capacity(Some("50"), Some("10")).unwrap_err();
-        assert!(
-            err.contains("OPERATOR_MAX_CAPACITY=50"),
-            "names chain value: {err}"
-        );
-        assert!(
-            err.contains("SANDBOX_MAX_COUNT=10"),
-            "names host value: {err}"
-        );
-    }
-
-    #[test]
-    fn capacity_cross_check_passes_equal_or_less() {
-        assert!(validate_chain_vs_host_capacity(Some("10"), Some("10")).is_ok());
-        assert!(validate_chain_vs_host_capacity(Some("5"), Some("10")).is_ok());
-    }
-
-    #[test]
-    fn capacity_cross_check_passes_uncapped_host() {
-        // SANDBOX_MAX_COUNT=0 disables the host cap entirely.
-        assert!(validate_chain_vs_host_capacity(Some("500"), Some("0")).is_ok());
-    }
-
-    #[test]
-    fn capacity_cross_check_requires_both_values() {
-        assert!(validate_chain_vs_host_capacity(None, Some("10")).is_ok());
-        assert!(validate_chain_vs_host_capacity(Some("50"), None).is_ok());
-        assert!(validate_chain_vs_host_capacity(None, None).is_ok());
-    }
-
-    #[test]
-    fn capacity_cross_check_ignores_unparseable_values() {
-        // Parity with the consumers: registration skips a bad
-        // OPERATOR_MAX_CAPACITY; the runtime defaults a bad SANDBOX_MAX_COUNT.
-        assert!(validate_chain_vs_host_capacity(Some("abc"), Some("10")).is_ok());
-        assert!(validate_chain_vs_host_capacity(Some("50"), Some("abc")).is_ok());
-    }
-}
+mod tests;
