+++ scripts/.god-object-baseline.txt
@@ -3,7 +3,6 @@
 # Doctrine: CLAUDE.md 'Module Size Discipline' — split by responsibility toward <600 LOC.
 2966 ui/src/routes/create.tsx
 1281 ui/src/routes/sandboxes.$id.tsx
-1111 ai-agent-sandbox-blueprint-bin/src/main.rs
 1106 ui/src/routes/instances.$id.tsx
 1095 ai-agent-sandbox-blueprint-lib/src/workflows.rs
 1065 sandbox-runtime/src/session_auth.rs
